+++ CHANGELOG.md
@@ -8,11 +8,14 @@ The format is based on [keep a changelog](http://keepachangelog.com/) and this p
 - Cursors can now be serialized and restored.
 - New storage partial update feature.
 - New storage list feature.
+- A new Unity code example which shows how to dispatch actions on the main thread.
+- A session now exposes `.ExpiresAt` and `.Handle` from the token.
 
 ### Changed
 - Added default builder for Notification listing and removal.
 - Group self list operations now return the user's membership type to each group.
 - Group leave operations now return a specific error code when the last admin attempts to leave.
+- The client interface now uses action delegates instead of event handlers to support a proxy pattern.
 
 ## [0.7.0] - 2017-07-18
 ### Added
@@ -26,7 +29,7 @@ The format is based on [keep a changelog](http://keepachangelog.com/) and this p
 
 ### Fixed
 - Accept SSL certificates.
-- Improve handling transport errors. 
+- Improve handling transport errors.
 - Improve fetching global storage records.
 
 ## [0.6.1] - 2017-05-30